shervn-backend/requirements.txt
python-dotenv
django 
djangorestframework 
django-cors-headers
psycopg2
Pillow